feat: code refactored
