+++ src/components/pages/categories/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import "./style.scss";
 
 import { Button, Select, Input } from "antd";
-import { DownloadOutlined } from "@ant-design/icons";
+import { PlusCircleOutlined } from "@ant-design/icons";
 
 const Categories = () => {
   const { Option } = Select;
@@ -16,42 +16,42 @@ const Categories = () => {
   }
 
   return (
-    <div className="page categories">
-      <div className="page__header">
+    <div className='page categories'>
+      <div className='page__header'>
         <h2>Categories</h2>
-        <Button type="primary" icon={<DownloadOutlined />}>
-          Download
+        <Button type='primary' icon={<PlusCircleOutlined />}>
+          Add New
         </Button>
       </div>
-      <div className="page__content">
-        <div className="row">
+      <div className='page__content'>
+        <div className='row'>
           <div>
             Show
             <Select
-              defaultValue="10"
+              defaultValue='10'
               style={{ width: 120, margin: "0 5px" }}
               onChange={handleChange}
             >
-              <Option value="25">25</Option>
-              <Option value="50">50</Option>
-              <Option value="100">100</Option>
+              <Option value='25'>25</Option>
+              <Option value='50'>50</Option>
+              <Option value='100'>100</Option>
             </Select>
             entries
           </div>
 
           <div>
             <Search
-              placeholder="search category"
+              placeholder='search category'
               allowClear
-              enterButton="Search"
-              size="large"
+              enterButton='Search'
+              size='large'
               onSearch={onSearch}
             />
           </div>
         </div>
 
         {/* Table */}
-        <div className="table"></div>
+        <div className='table'></div>
         {/* Table */}
       </div>
     </div>
