+++ src/styles/_normalize.scss
@@ -5,4 +5,5 @@ h4,
 h5,
 h6 {
   margin: 0;
+  padding: 0;
 }
