+++ src/styles/_typography.scss
@@ -1,0 +1,7 @@
+body {
+  font-family: "Poppins", sans-serif;
+  font-weight: 400;
+  line-height: 1.7;
+  font-size: $default-font-size;
+  color: $color-black;
+}
