+++ src/styles/_utilities.scss
@@ -1,0 +1,3 @@
+.u-width-100 {
+  width: 100%;
+}
