+++ src/styles/index.scss
@@ -1,2 +1,9 @@
-@import "~styles/global";
-@import "~styles/normalize";
+@import "~antd/dist/antd.css";
+
+@import "_base";
+@import "normalize.scss";
+@import "variables.scss";
+@import "mixins.scss";
+@import "typography.scss";
+@import "utilities.scss";
+@import "global.scss";
